Extract message-list scrolling into useMessageListScroll
Pure relocation: live-edge following (use-stick-to-bottom plus the
gesture-attribution layer), the new-turn reading-line reserve, and
windowed history rendering move from message-list.tsx into a colocated
hook. The component keeps rendering; the hook decides where the
viewport is. No behavior change — the 120-test message-list suite and
the follow E2E specs pass unmodified.
